# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bomato restaurant listing API
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: number;
  name: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  cuisines: string[];
  /** 1 to 4 dollar signs */
  priceLevel: number;
  address: string;
  /**
     * Estimated delivery time in minutes
     * @nullable
     */
  deliveryTime?: number | null;
  isOpen: boolean;
  /** @nullable */
  isFeatured?: boolean | null;
}